refactor(daemon): extract session orchestrator
Why:
- Move the Session lifecycle brain out of the FastAPI WebSocket adapter so lifecycle tests can exercise Daemon behavior through EventSink without transport fakes.
- Preserve the existing WebSocket wire protocol while letting server.py remain a thin adapter around parsed client messages.

Side effects:
- Session guard sample-limit trips now finalize through the same auto-stop path as the wall-clock limit.
- Adds RecordingEventSink orchestrator coverage for busy, missing, abort, disconnect, limit, and finalization flows.
