+++ parakeet-stt-daemon/tests/test_overlay_event_stream.py
@@ -13,16 +13,19 @@ from uuid import UUID, uuid4
 import numpy as np
 from loguru import logger
 
-from parakeet_stt_daemon import server as server_module
+from parakeet_stt_daemon import session_orchestrator as orchestrator_module
 from parakeet_stt_daemon.config import ServerSettings
+from parakeet_stt_daemon.events import WebSocketEventSinkState
 from parakeet_stt_daemon.messages import (
     AbortSession,
     ClientMessageType,
+    ParsedMessage,
     StartSession,
     StopSession,
 )
 from parakeet_stt_daemon.server import DaemonServer
 from parakeet_stt_daemon.session import SessionManager
+from parakeet_stt_daemon.session_orchestrator import SessionOrchestrator
 
 
 class FakeAudio:
@@ -44,6 +47,9 @@ class FakeAudio:
 
     def abort_session(self) -> None:
         self.abort_calls += 1
+
+    def take_audio_levels(self) -> list[float]:
+        return []
 
     def take_stream_chunks(self) -> list[np.ndarray]:
         return []
@@ -139,51 +145,77 @@ def _build_server(
     incremental_fail_at_call: int | None = None,
 ) -> DaemonServer:
     server = cast(Any, DaemonServer.__new__(DaemonServer))
-    server.settings = ServerSettings(
+    orchestrator = cast(Any, SessionOrchestrator.__new__(SessionOrchestrator))
+    settings = ServerSettings(
         device="cpu",
         status_enabled=True,
         streaming_enabled=False,
         overlay_events_enabled=overlay_events_enabled,
     )
-    server.sessions = SessionManager()
-    server.audio = FakeAudio(ready_chunks=ready_chunks)
-    server.model = object()
-    server.transcriber = FakeIncrementalTranscriber(
+    server.settings = settings
+    server.orchestrator = orchestrator
+    server.event_sinks = WebSocketEventSinkState(
+        overlay_events_enabled=settings.overlay_events_enabled
+    )
+    server._websocket_send_locks = {}
+    orchestrator.settings = settings
+    orchestrator.sessions = SessionManager()
+    orchestrator.audio = FakeAudio(ready_chunks=ready_chunks)
+    orchestrator.model = object()
+    orchestrator.transcriber = FakeIncrementalTranscriber(
         incremental_outputs,
         fail_at_call=incremental_fail_at_call,
     )
-    server._session_lock = asyncio.Lock()
-    server._inference_lock = asyncio.Lock()
-    server.streaming_transcriber = None
-    server._active_stream = None
-    server._stream_drain_task = None
-    server._stream_drain_running = False
-    server._session_guard_task = None
-    server._session_guard_running = False
-    server._session_sample_limit = 1_600
-    server._session_age_limit_ms = 90_000
-    server._requested_device = "cpu"
-    server._effective_device = "cpu"
-    server._last_audio_ms = None
-    server._last_audio_stop_ms = None
-    server._last_finalize_ms = None
-    server._last_infer_ms = None
-    server._last_send_ms = None
-    server._live_interim_audio = np.zeros((0,), dtype=np.float32)
-    server._live_interim_failed = False
-    server._overlay_event_seq_by_session = {}
-    server._overlay_last_interim_text_by_session = {}
-    server._overlay_interim_stabilizer_by_session = {}
-    server._overlay_state_by_session = {}
-    server._overlay_events_emitted = 0
-    server._overlay_events_dropped = 0
-    server._websocket_send_locks = {}
+    orchestrator._session_lock = asyncio.Lock()
+    orchestrator._inference_lock = asyncio.Lock()
+    orchestrator.streaming_transcriber = None
+    orchestrator._active_stream = None
+    orchestrator._stream_drain_task = None
+    orchestrator._stream_drain_running = False
+    orchestrator._session_guard_task = None
+    orchestrator._session_guard_running = False
+    orchestrator._session_sample_limit = 1_600
+    orchestrator._session_age_limit_ms = 90_000
+    orchestrator._requested_device = "cpu"
+    orchestrator._effective_device = "cpu"
+    orchestrator._last_audio_ms = None
+    orchestrator._last_audio_stop_ms = None
+    orchestrator._last_finalize_ms = None
+    orchestrator._last_infer_ms = None
+    orchestrator._last_send_ms = None
+    orchestrator._live_interim_audio = np.zeros((0,), dtype=np.float32)
+    orchestrator._live_interim_failed = False
+    orchestrator._overlay_interim_stabilizer_by_session = {}
 
     async def fake_finalize(_audio_samples: np.ndarray) -> tuple[str, int]:
         return "overlay test text", 7
 
-    server._finalise_transcription = fake_finalize
+    orchestrator._finalise_transcription = fake_finalize
     return cast(DaemonServer, server)
+
+
+async def _dispatch_message(
+    server: DaemonServer,
+    websocket: FakeWebSocket,
+    message: StartSession | StopSession | AbortSession,
+) -> None:
+    await server._dispatch(
+        cast(Any, websocket),
+        ParsedMessage(kind=ClientMessageType(message.type), model=message),
+    )
+
+
+async def _emit_live_interim_from_chunk(
+    server: DaemonServer,
+    websocket: FakeWebSocket,
+    session_id: UUID,
+    chunk: np.ndarray,
+) -> None:
+    await server.orchestrator._emit_live_interim_from_chunk(
+        server._event_sink(cast(Any, websocket)),
+        session_id,
+        chunk,
+    )
 
 
 def _start_message(session_id: UUID) -> StartSession:
@@ -206,7 +238,7 @@ def _disable_server_sleep(monkeypatch) -> None:
     async def no_sleep(_seconds: float) -> None:
         return None
 
-    monkeypatch.setattr("parakeet_stt_daemon.server.asyncio.sleep", no_sleep)
+    monkeypatch.setattr("parakeet_stt_daemon.session_orchestrator.asyncio.sleep", no_sleep)
 
 
 def _pause_guard_sleep(monkeypatch) -> tuple[asyncio.Event, asyncio.Event]:
@@ -217,7 +249,7 @@ def _pause_guard_sleep(monkeypatch) -> tuple[asyncio.Event, asyncio.Event]:
         guard_sleep_entered.set()
         await allow_guard_resume.wait()
 
-    monkeypatch.setattr(server_module, "_REAL_ASYNCIO_SLEEP", fake_guard_sleep)
+    monkeypatch.setattr(orchestrator_module, "_REAL_ASYNCIO_SLEEP", fake_guard_sleep)
     return guard_sleep_entered, allow_guard_resume
 
 
@@ -239,8 +271,8 @@ def test_overlay_events_disabled_emits_only_baseline_messages(monkeypatch) -> No
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types == ["session_started", "final_result"]
@@ -255,15 +287,15 @@ def test_stop_session_finalizes_gracefully_when_sample_limit_breached(monkeypatc
         guard_sleep_entered, allow_guard_resume = _pause_guard_sleep(monkeypatch)
 
         server = _build_server(overlay_events_enabled=False)
-        audio = cast(FakeAudio, server.audio)
+        audio = cast(FakeAudio, server.orchestrator.audio)
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
         await asyncio.wait_for(guard_sleep_entered.wait(), timeout=1.0)
 
         audio._session_limit_exceeded = True
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
         allow_guard_resume.set()
         await asyncio.sleep(0)
 
@@ -272,7 +304,7 @@ def test_stop_session_finalizes_gracefully_when_sample_limit_breached(monkeypatc
         assert "final_result" in sent_types
         assert audio.abort_calls == 0
         assert audio.stop_calls == 1
-        assert server.sessions.active is None
+        assert server.orchestrator.sessions.active is None
 
     asyncio.run(scenario())
 
@@ -285,8 +317,8 @@ def test_overlay_events_enabled_stream_is_ordered_and_final_once(monkeypatch) ->
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types == [
@@ -330,10 +362,10 @@ def test_overlay_sequence_does_not_leak_across_sessions(monkeypatch) -> None:
         first = uuid4()
         second = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(first))
-        await server._handle_stop(cast(Any, websocket), _stop_message(first))
-        await server._handle_start(cast(Any, websocket), _start_message(second))
-        await server._handle_stop(cast(Any, websocket), _stop_message(second))
+        await _dispatch_message(server, websocket, _start_message(first))
+        await _dispatch_message(server, websocket, _stop_message(first))
+        await _dispatch_message(server, websocket, _start_message(second))
+        await _dispatch_message(server, websocket, _stop_message(second))
 
         interim_by_session: dict[str, list[int]] = {}
         for payload in websocket.sent_json:
@@ -364,8 +396,8 @@ def test_overlay_send_failures_do_not_block_final_result(monkeypatch) -> None:
 
         _set_dynamic_attr(websocket, "send_json", send_json)
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types == ["session_started", "final_result"]
@@ -394,8 +426,8 @@ def test_interim_text_emitted_when_incremental_source_validates(monkeypatch) -> 
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types == [
@@ -442,8 +474,8 @@ def test_incremental_source_failure_does_not_break_final_result(monkeypatch) -> 
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types == [
@@ -474,18 +506,18 @@ def test_live_interim_chunk_emission_confirms_repeated_text_before_append(monkey
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.3, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.3, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.4, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.4, dtype=np.float32)
         )
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
@@ -521,9 +553,9 @@ def test_live_interim_first_snapshot_is_visible_immediately(monkeypatch) -> None
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
         )
 
         interim_texts = [
@@ -547,15 +579,15 @@ def test_live_interim_zero_overlap_rewrite_replaces_mutable_tail(monkeypatch) ->
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.3, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.3, dtype=np.float32)
         )
 
         interim_texts = [
@@ -581,12 +613,12 @@ def test_live_interim_stabilizer_logs_when_streaming_debug_enabled(monkeypatch) 
         session_id = uuid4()
 
         with _capture_loguru_messages() as log_output:
-            await server._handle_start(cast(Any, websocket), _start_message(session_id))
-            await server._emit_live_interim_from_chunk(
-                cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+            await _dispatch_message(server, websocket, _start_message(session_id))
+            await _emit_live_interim_from_chunk(
+                server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
             )
-            await server._emit_live_interim_from_chunk(
-                cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
+            await _emit_live_interim_from_chunk(
+                server, cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
             )
 
         messages = log_output.getvalue()
@@ -616,8 +648,8 @@ def test_stop_path_stabilizer_logs_when_streaming_debug_enabled(monkeypatch) -> 
         session_id = uuid4()
 
         with _capture_loguru_messages() as log_output:
-            await server._handle_start(cast(Any, websocket), _start_message(session_id))
-            await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+            await _dispatch_message(server, websocket, _start_message(session_id))
+            await _dispatch_message(server, websocket, _stop_message(session_id))
 
         messages = log_output.getvalue()
         assert "source=stop_replay source_seq=0" in messages
@@ -642,9 +674,9 @@ def test_stabilizer_logs_empty_candidate_action_when_debug_enabled(monkeypatch) 
         session_id = uuid4()
 
         with _capture_loguru_messages() as log_output:
-            await server._handle_start(cast(Any, websocket), _start_message(session_id))
-            await server._emit_live_interim_from_chunk(
-                cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+            await _dispatch_message(server, websocket, _start_message(session_id))
+            await _emit_live_interim_from_chunk(
+                server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
             )
 
         messages = log_output.getvalue()
@@ -670,9 +702,9 @@ def test_stabilizer_logs_skip_on_live_transcribe_error_when_debug_enabled(monkey
         session_id = uuid4()
 
         with _capture_loguru_messages() as log_output:
-            await server._handle_start(cast(Any, websocket), _start_message(session_id))
-            await server._emit_live_interim_from_chunk(
-                cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+            await _dispatch_message(server, websocket, _start_message(session_id))
+            await _emit_live_interim_from_chunk(
+                server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
             )
 
         messages = log_output.getvalue()
@@ -697,12 +729,12 @@ def test_stabilizer_logging_is_disabled_by_default(monkeypatch) -> None:
         session_id = uuid4()
 
         with _capture_loguru_messages() as log_output:
-            await server._handle_start(cast(Any, websocket), _start_message(session_id))
-            await server._emit_live_interim_from_chunk(
-                cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+            await _dispatch_message(server, websocket, _start_message(session_id))
+            await _emit_live_interim_from_chunk(
+                server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
             )
-            await server._emit_live_interim_from_chunk(
-                cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
+            await _emit_live_interim_from_chunk(
+                server, cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
             )
 
         messages = log_output.getvalue()
@@ -723,12 +755,12 @@ def test_live_interim_case_only_updates_are_emitted(monkeypatch) -> None:
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.1, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
         )
 
         interim_texts = [
@@ -763,10 +795,10 @@ def test_live_interim_stabilizer_preserves_full_session_transcript_when_window_r
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
         for chunk in ready_chunks[:3]:
-            await server._emit_live_interim_from_chunk(cast(Any, websocket), session_id, chunk)
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+            await _emit_live_interim_from_chunk(server, websocket, session_id, chunk)
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         interim_texts = [
             cast(str, payload["text"])
@@ -790,7 +822,7 @@ def test_stop_waits_for_in_flight_live_interim_before_final_send(monkeypatch) ->
 
         server = _build_server(overlay_events_enabled=True)
         transcriber = BlockingIncrementalTranscriber("late interim")
-        _set_dynamic_attr(server, "transcriber", transcriber)
+        _set_dynamic_attr(server.orchestrator, "transcriber", transcriber)
         websocket = FakeWebSocket()
         session_id = uuid4()
         final_started = asyncio.Event()
@@ -804,20 +836,21 @@ def test_stop_waits_for_in_flight_live_interim_before_final_send(monkeypatch) ->
 
         _set_dynamic_attr(websocket, "send_json", send_json)
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
         live_task = asyncio.create_task(
-            server._emit_live_interim_from_chunk(
+            _emit_live_interim_from_chunk(
+                server,
                 cast(Any, websocket),
                 session_id,
                 np.full((400,), 0.2, dtype=np.float32),
             )
         )
-        server._stream_drain_task = live_task
-        server._stream_drain_running = True
+        server.orchestrator._stream_drain_task = live_task
+        server.orchestrator._stream_drain_running = True
         await asyncio.to_thread(transcriber.started.wait)
 
         stop_task = asyncio.create_task(
-            server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+            _dispatch_message(server, websocket, _stop_message(session_id))
         )
         await asyncio.sleep(0)
 
@@ -855,29 +888,33 @@ def test_stop_path_serializes_live_interim_and_final_decode(monkeypatch) -> None
 
         server = _build_server(overlay_events_enabled=True)
         transcriber = SerializingTranscriber()
-        _set_dynamic_attr(server, "transcriber", transcriber)
+        _set_dynamic_attr(server.orchestrator, "transcriber", transcriber)
         _set_dynamic_attr(
-            server,
+            server.orchestrator,
             "_finalise_transcription",
-            DaemonServer._finalise_transcription.__get__(server, DaemonServer),
+            SessionOrchestrator._finalise_transcription.__get__(
+                server.orchestrator,
+                SessionOrchestrator,
+            ),
         )
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
         live_task = asyncio.create_task(
-            server._emit_live_interim_from_chunk(
+            _emit_live_interim_from_chunk(
+                server,
                 cast(Any, websocket),
                 session_id,
                 np.full((400,), 0.2, dtype=np.float32),
             )
         )
-        server._stream_drain_task = live_task
-        server._stream_drain_running = True
+        server.orchestrator._stream_drain_task = live_task
+        server.orchestrator._stream_drain_running = True
         await asyncio.to_thread(transcriber.live_started.wait)
 
         stop_task = asyncio.create_task(
-            server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+            _dispatch_message(server, websocket, _stop_message(session_id))
         )
         await asyncio.sleep(0)
 
@@ -909,9 +946,10 @@ def test_abort_emits_session_ended_and_no_final_result() -> None:
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_abort(
-            cast(Any, websocket),
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(
+            server,
+            websocket,
             AbortSession(
                 type=ClientMessageType.ABORT_SESSION,
                 session_id=session_id,
@@ -943,8 +981,8 @@ def test_phase6_quick_utterance_contract_preserves_final_once(monkeypatch) -> No
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types.count("final_result") == 1
@@ -994,10 +1032,10 @@ def test_phase6_long_dictation_contract_preserves_monotonic_interim_tail(monkeyp
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
         for chunk in ready_chunks:
-            await server._emit_live_interim_from_chunk(cast(Any, websocket), session_id, chunk)
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+            await _emit_live_interim_from_chunk(server, websocket, session_id, chunk)
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types.count("final_result") == 1
@@ -1037,22 +1075,22 @@ def test_phase6_daemon_reconnect_contract_recovers_with_fresh_session(monkeypatc
         first_session = uuid4()
         second_session = uuid4()
 
-        await server._handle_start(cast(Any, first_socket), _start_message(first_session))
-        await server._emit_live_interim_from_chunk(
-            cast(Any, first_socket), first_session, np.full((400,), 0.2, dtype=np.float32)
+        await _dispatch_message(server, first_socket, _start_message(first_session))
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, first_socket), first_session, np.full((400,), 0.2, dtype=np.float32)
         )
-        await server._emit_live_interim_from_chunk(
-            cast(Any, first_socket), first_session, np.full((400,), 0.3, dtype=np.float32)
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, first_socket), first_session, np.full((400,), 0.3, dtype=np.float32)
         )
-        cleaned = await server._cleanup_active_session(
+        cleaned = await server.orchestrator._cleanup_active_session(
             "websocket disconnected",
             expected_session_id=first_session,
             require_session_match=True,
         )
         assert cleaned is True
-        assert server.sessions.active is None
-        await server._handle_start(cast(Any, second_socket), _start_message(second_session))
-        await server._handle_stop(cast(Any, second_socket), _stop_message(second_session))
+        assert server.orchestrator.sessions.active is None
+        await _dispatch_message(server, second_socket, _start_message(second_session))
+        await _dispatch_message(server, second_socket, _stop_message(second_session))
 
         sent_types = [cast(str, payload["type"]) for payload in second_socket.sent_json]
         assert sent_types.count("final_result") == 1
@@ -1089,11 +1127,11 @@ def test_phase6_overlay_crash_mid_session_contract_keeps_final_non_fatal(monkeyp
 
         _set_dynamic_attr(websocket, "send_json", flaky_send_json)
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._emit_live_interim_from_chunk(
-            cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _emit_live_interim_from_chunk(
+            server, cast(Any, websocket), session_id, np.full((400,), 0.2, dtype=np.float32)
         )
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
         sent_types = [cast(str, payload["type"]) for payload in websocket.sent_json]
         assert sent_types.count("final_result") == 1
@@ -1115,12 +1153,12 @@ def test_live_interim_context_window_grows_across_session() -> None:
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
         chunk = np.full((8_000,), 0.2, dtype=np.float32)
         for _ in range(8):
-            await server._emit_live_interim_from_chunk(cast(Any, websocket), session_id, chunk)
+            await _emit_live_interim_from_chunk(server, websocket, session_id, chunk)
 
-        sample_sizes = cast(Any, server.transcriber).sample_sizes
+        sample_sizes = cast(Any, server.orchestrator.transcriber).sample_sizes
         assert sample_sizes == sorted(sample_sizes)
         assert sample_sizes[-1] == 8 * 8_000
 
@@ -1140,10 +1178,10 @@ def test_stop_path_interim_context_window_grows_across_chunks(monkeypatch) -> No
         websocket = FakeWebSocket()
         session_id = uuid4()
 
-        await server._handle_start(cast(Any, websocket), _start_message(session_id))
-        await server._handle_stop(cast(Any, websocket), _stop_message(session_id))
+        await _dispatch_message(server, websocket, _start_message(session_id))
+        await _dispatch_message(server, websocket, _stop_message(session_id))
 
-        sample_sizes = cast(Any, server.transcriber).sample_sizes
+        sample_sizes = cast(Any, server.orchestrator.transcriber).sample_sizes
         assert sample_sizes == sorted(sample_sizes)
         assert sample_sizes[-1] == 8 * 8_000
